requirements.txt
request
pandas>=1.3
python-dotenv
lxml
tabulate
numpy